gemini3-zkp-attestation-agent/requirements.txt
# ============================================
# WEB FRAMEWORK
# ============================================
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0

# ============================================
# DATABASE
# ============================================
sqlalchemy==2.0.25
alembic==1.13.1
asyncpg==0.29.0

# ============================================
# CRYPTOGRAPHY & ZKP
# ============================================
cryptography==42.0.0
pycryptodome==3.19.1

# ============================================
# ALGORAND (ON-CHAIN)
# ============================================
py-algorand-sdk==2.6.1
pyteal==0.27.0

# ============================================
# HTTP & ASYNC
# ============================================
httpx==0.26.0
aiofiles==23.2.1
requests==2.31.0

# ============================================
# SECURITY
# ============================================
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-dotenv==1.0.0

# ============================================
# UTILITIES
# ============================================
python-json-logger==2.0.7

# ============================================
# MONITORING
# ============================================
prometheus-client==0.19.0

# ============================================
# BACKGROUND TASKS
# ============================================
celery==5.3.4
redis==5.0.1

# ============================================
# TESTING
# ============================================
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
pytest-mock==3.12.0
